DcSwitchStateMachine: support RILs not implementing REQUEST_ALLOW_DATA
When a RIL returns REQUEST_NOT_SUPPORTED in response to
RIL_REQUEST_ALLOW_DATA this is seen as an error and no state
transition to attached is carried out when the modem is already
attached. Fix this by handling REQUEST_NOT_SUPPORTED as success.

This fixes devices randomly not attaching data after reboot if
the modem returns attached in DATA_REGISTRATION_STATE before
setDataAllowed is called.

Change-Id: I1e58be9d8a8bfe67df4c266e5670a47acedf6f03
Signed-off-by: Raleigh Matlock <raleighman2@gmail.com>